Put parts of HomeView class into new classes
