+++ infrastructure/submission_site/lib/home/view/home_view.dart
@@ -1,10 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:moralpainapi/moralpainapi.dart';
-import 'package:submission_site/datetime/datetime.dart';
 import 'package:submission_site/home/home.dart';
-import 'package:submission_site/score/score.dart';
-import 'package:submission_site/selections/selections.dart';
 
 class HomeView extends StatelessWidget {
   const HomeView({Key? key}) : super(key: key);
@@ -13,6 +9,7 @@ class HomeView extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocBuilder<HomeBloc, HomeState>(
       builder: (context, state) {
+        print('rebuilding HomeView!');
         switch (state.submissionStatus) {
           case SubmissionStatus.initial:
           case SubmissionStatus.loading:
@@ -27,7 +24,6 @@ class HomeView extends StatelessWidget {
                   body: Center(child: CircularProgressIndicator()),
                 );
               case SurveyStatus.success:
-                final bloc = BlocProvider.of<HomeBloc>(context);
                 final areChanges =
                     state.timestamp != state.submission!.timestamp ||
                         state.score != state.submission!.score ||
@@ -49,132 +45,9 @@ class HomeView extends StatelessWidget {
                               crossAxisAlignment: CrossAxisAlignment.center,
                               mainAxisAlignment: MainAxisAlignment.center,
                               children: [
-                                FieldDisplay(
-                                  text:
-                                      'Time submitted: ${displayTimestamp(state.timestamp!)}',
-                                  onPressed: () {
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                        builder: (context) =>
-                                            BlocProvider<HomeBloc>.value(
-                                          value: bloc,
-                                          child: const DatetimeRoute(),
-                                        ),
-                                        fullscreenDialog: true,
-                                      ),
-                                    );
-                                  },
-                                ),
-                                FieldDisplay(
-                                  text: 'Score: ${state.score} out of 10',
-                                  onPressed: () {
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                        builder: (context) =>
-                                            BlocProvider<HomeBloc>.value(
-                                          value: bloc,
-                                          child: const ScoreRoute(),
-                                        ),
-                                        fullscreenDialog: true,
-                                      ),
-                                    );
-                                  },
-                                ),
-                                FieldDisplay(
-                                  text: 'Contributing factors: '
-                                      '${displaySelections(state.selections!, state.survey!)}',
-                                  onPressed: () {
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                        builder: (context) =>
-                                            BlocProvider<HomeBloc>.value(
-                                          value: bloc,
-                                          child: const SelectionsRoute(),
-                                        ),
-                                        fullscreenDialog: true,
-                                      ),
-                                    );
-                                  },
-                                ),
+                                FieldList(),
                                 const SizedBox(height: 50.0),
-                                Row(
-                                  mainAxisAlignment: MainAxisAlignment.center,
-                                  children: [
-                                    ElevatedButton(
-                                      onPressed: areChanges
-                                          ? () {
-                                              final alert = AlertDialog(
-                                                title: const Text(
-                                                  'Are you sure you want to revert your changes?',
-                                                ),
-                                                actions: [
-                                                  TextButton(
-                                                    onPressed: () {
-                                                      Navigator.of(context)
-                                                          .pop();
-                                                    },
-                                                    child: const Text('No'),
-                                                  ),
-                                                  TextButton(
-                                                    onPressed: () {
-                                                      bloc.add(
-                                                        const HomeChangesReverted(),
-                                                      );
-                                                      Navigator.of(context)
-                                                          .pop();
-                                                    },
-                                                    child: const Text('Yes'),
-                                                  )
-                                                ],
-                                              );
-                                              showDialog(
-                                                  context: context,
-                                                  builder: (_) => alert);
-                                            }
-                                          : null,
-                                      child: const Text('Revert Changes'),
-                                    ),
-                                    const SizedBox(width: 30.0),
-                                    ElevatedButton(
-                                      onPressed: areChanges
-                                          ? () {
-                                              final alert = AlertDialog(
-                                                title: const Text(
-                                                  'Are you sure you want to submit your changes?',
-                                                ),
-                                                actions: [
-                                                  TextButton(
-                                                    onPressed: () {
-                                                      Navigator.of(context)
-                                                          .pop();
-                                                    },
-                                                    child: const Text('No'),
-                                                  ),
-                                                  TextButton(
-                                                    onPressed: () async {
-                                                      bloc.add(
-                                                        const HomeChangesSubmitted(),
-                                                      );
-                                                      Navigator.of(context)
-                                                          .pop();
-                                                    },
-                                                    child: const Text('Yes'),
-                                                  )
-                                                ],
-                                              );
-                                              showDialog(
-                                                  context: context,
-                                                  builder: (_) => alert);
-                                            }
-                                          : null,
-                                      child: const Text('Submit Changes'),
-                                    ),
-                                    const SizedBox(height: 50.0)
-                                  ],
-                                )
+                                RevertSubmitChangesWidget(enabled: areChanges),
                               ],
                             ),
                           ),
@@ -209,42 +82,6 @@ class HomeView extends StatelessWidget {
       if (a[i] != b[i]) return false;
     }
     return true;
-  }
-
-  @visibleForTesting
-  static String displayTimestamp(int timestamp) {
-    final datetime =
-        DateTime.fromMillisecondsSinceEpoch(timestamp * 1000).toUtc();
-    String datetimeString = datetime.toString();
-    datetimeString = datetimeString.substring(0, datetimeString.length - 5);
-    datetimeString += ' (UTC)';
-    return datetimeString;
-  }
-
-  @visibleForTesting
-  static String displaySelections(List<String> selections, Survey survey) {
-    if (selections.isEmpty) return ' none';
-    String ret = '\n';
-    for (String id in selections) {
-      ret += '• ${getSelectionDescription(id, survey)}';
-      if (id != selections.last) ret += '\n';
-    }
-    return ret;
-  }
-
-  @visibleForTesting
-  static String getSelectionDescription(String id, Survey survey) {
-    List<String> ids = id.split('_');
-    for (SurveySection section in survey.sections!) {
-      if (section.sectionId! == ids[0]) {
-        for (SurveyOption option in section.options!) {
-          if (option.id! == ids[1]) {
-            return option.description!;
-          }
-        }
-      }
-    }
-    return 'Couldn\'t find description.';
   }
 
   @visibleForTesting
